Fix all Java coding standard issues
