+++ src/main/java/Commands/Command.java
@@ -1,8 +1,8 @@
 package Commands;
 
-import TaskList.TaskList;
-import Ui.Ui;
-import Storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+import storage.Storage;
 
 public interface Command {
     void execute(TaskList tasks, Ui ui, Storage storage);
