+++ src/main/java/storage/Storage.java
@@ -1,0 +1,31 @@
+package storage;
+
+import java.util.ArrayList;
+import tasklist.TaskList;
+import tasks.Task;
+
+public class Storage {
+  private static String path;
+
+  private DataReader dataReader;
+
+  private DataWriter dataWriter;
+
+  public Storage(String path) {
+    Storage.path = path;
+    this.dataReader = new DataReader();
+    this.dataWriter = new DataWriter(path);
+  }
+
+  public static void addLine(String line) {
+    DataWriter.addLine(line);
+  }
+
+  public static void refresh(TaskList taskList) {
+    DataWriter.refresh(taskList);
+  }
+
+  public static ArrayList<Task> readFileToTasksLists(String fileName) {
+    return DataReader.readTasksFromFile(fileName);
+  }
+}
