+++ src/main/java/tasks/Task.java
@@ -1,0 +1,33 @@
+package tasks;
+
+public class Task {
+  protected String description;
+  protected boolean isDone;
+
+  public Task(String description) {
+    this.description = description;
+    this.isDone = false;
+  }
+
+  public Task(String des, boolean mark) {
+    this.description = des;
+    this.isDone = mark;
+  }
+
+  public String getStatusIcon() {
+    return (isDone ? "[X]" : "[ ]"); // mark done task with X
+  }
+
+  public void toMark() {
+    this.isDone = true;
+  }
+
+  public void toUnmark() {
+    this.isDone = false;
+  }
+
+  @Override
+  public String toString() {
+    return this.getStatusIcon() + " " + description;
+  }
+}
